requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
django-environ==0.9.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
django-import-export     ==     3.0.2

idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2 == 2.9.5
psycopg2-binary==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.6
phonenumbers==8.13.0
python-http-client==3.3.7
pytz==2022.6
requests==2.28.1
# ruamel.yaml==0.17.21
# ruamel.yaml.clib==0.2.7
sendgrid==6.9.7
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
uritemplate==4.1.1
urllib3==1.26.12
pandas
numpy==1.23.4
openpyxl==3.0.10
django-filter==22.1
django-cors-headers
pillow
django-import-export
openpyxl
xlwt
cryptography
paramiko
fernet 